requirements.txt
pandas
numpy<2.0
evidently==0.4.9
psycopg2-binary
pydantic==1.10.13
